+++ griptape/drivers/web_search/tavily_web_search_driver.py
@@ -3,8 +3,9 @@ from __future__ import annotations
 from typing import TYPE_CHECKING, Any
 
 from attrs import Factory, define, field
+from tavily import BadRequestError
 
-from griptape.artifacts import JsonArtifact, ListArtifact, TextArtifact
+from griptape.artifacts import JsonArtifact, ListArtifact
 from griptape.drivers import BaseWebSearchDriver
 from griptape.utils import import_optional_dependency
 
@@ -23,5 +24,7 @@ class TavilyWebSearchDriver(BaseWebSearchDriver):
 
     def search(self, query: str, **kwargs) -> ListArtifact:
         response = self.client.search(query, max_results=self.results_count, **self.params, **kwargs)
-        results = response["results"]
-        return ListArtifact([TextArtifact(JsonArtifact(result)) for result in results])
+        results = response.get("results", [])
+        if not results:
+            raise BadRequestError("No results found or the response structure is invalid.")
+        return ListArtifact([(JsonArtifact(result)) for result in results])